Replaced unsigned by size_t, repaired unix-script
git-svn-id: https://icmake.svn.sourceforge.net/svnroot/icmake/icmake/trunk@41 dcc949cb-2a1c-0410-b67a-e7e5ca830264

+++ icmake/comp/addpatch.c
@@ -4,9 +4,9 @@
 
 #include "iccomp.h"
 
-void addpatch (unsigned *list, unsigned len, register unsigned value)
+void addpatch (size_t *list, size_t len, register size_t value)
 {
-    register unsigned
+    register size_t
         index;
 
     for (index = 0; index < len; index++)
