+++ icmake/comp/patchtru.c
@@ -8,7 +8,7 @@ void patchtrue(e)
     ESTRUC_
         *e;
 {                                           /* expand the truelist */
-    e->truelist = xrealloc(e->truelist, (e->truelen + 1) * sizeof(unsigned));
+    e->truelist = xrealloc(e->truelist, (e->truelen + 1) * sizeof(size_t));
 
                                             /* room for the jump-backpatch */
     e->code = xrealloc(e->code, e->codelen += sizeof(INT16));
