+++ icmake/rss/redirect.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include "icrssdef.h"
 
-unsigned redirect_start (unsigned x, unsigned y)
+size_t redirect_start (size_t x, size_t y)
 {
     return (1);
 }
